写入MongoDB的数据库连接地址,如果设置有mongdb的用户名,密码则添加spring.data.mongodb.uri=mongodb://name:pass@localhost:27017/test到对应的配置文件中.

+++ parent/springboot/src/main/java/com/mongodb/springboot/dao/UserDao.java
@@ -9,7 +9,7 @@ import com.mongodb.springboot.entity.UserEntity;
 public interface UserDao  {
   void saveUser(UserEntity user);
   UserEntity findUserByUserName(String userName);
-  int updateUser(UserEntity user);
+  Long updateUser(UserEntity user);
   void deleteUserById(Long id);
 
 }
